# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for the owner-managed Project Hub catalog.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: Browse published projects
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: access
          schema: { type: string, enum: [all, free, premium] }
        - in: query
          name: sort
          schema: { type: string, enum: [latest, popular] }
        - in: query
          name: featured
          schema: { type: boolean }
      responses:
        "200":
          description: Matching published projects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: View a published project
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectDetail" }
        "404":
          description: Project not found
  /projects/{slug}/views:
    post:
      operationId: recordProjectView
      tags: [projects]
      summary: Record a project page view
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: View recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionResult" }
  /projects/{slug}/download:
    post:
      operationId: requestProjectDownload
      tags: [projects]
      summary: Request a download after server-side access checks
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Download request accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DownloadResult" }
        "401":
          description: Sign-in or an active premium entitlement is required
        "403":
          description: The project file is unavailable or access is not allowed
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: project_hub_session
  parameters:
    Slug:
      in: path
      name: slug
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
  schemas:
...
    Project:
      type: object
      required:
        - id
        - slug
        - title
        - shortDescription
        - description
        - category
        - tags
        - access
        - status
        - featured
        - isNew
        - version
        - views
        - downloads
        - favorites
        - releaseDate
        - updatedAt
        - screenshots
        - supportedDevices
        - requirements
        - whatsNew
        - changelog
        - fileAvailable
        - videoLocked
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        shortDescription: { type: string }
        description: { type: string }
        category: { type: string }
        tags:
          type: array
          items: { type: string }
        access: { type: string, enum: [free, premium] }
        status: { type: string, enum: [draft, published, archived] }
        featured: { type: boolean }
        isNew: { type: boolean }
        thumbnailUrl: { type: ["string", "null"] }
        previewVideoUrl: { type: ["string", "null"] }
        version: { type: string }
        fileSize: { type: ["string", "null"] }
        views: { type: integer }
        downloads: { type: integer }
        favorites: { type: integer }
        releaseDate: { type: string, format: date }
        updatedAt: { type: string, format: date-time }
        screenshots:
          type: array
          items: { type: string }
        supportedDevices:
          type: array
          items: { type: string }
        requirements:
          type: array
          items: { type: string }
        whatsNew: { type: string }
        changelog:
          type: array
          items: { type: string }
        fileAvailable: { type: boolean }
        videoLocked: { type: boolean }
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          required: [isFavorite]
          properties:
            isFavorite: { type: boolean }
...
    ActionResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: ["string", "null"] }
    DownloadResult:
      type: object
      required: [success, downloadUrl]
      properties:
        success: { type: boolean }
        downloadUrl: { type: ["string", "null"] }
        message: { type: ["string", "null"] }
...